testing with utf8
The test file needed to use utf8 in order to create strings that would match the regexes in the module; either they should both use utf8 or neither should.

Also added the degree sign and a local variable stanza for emacs to know what to do with the test file without cperl owning the .t extension.